Allow building on more platforms
This extends changes made in #538, but with more targeted exclusions
making more of the API surface area available to clients.

+++ lib/Core/SQLiteBuildDB.cpp
@@ -548,7 +548,7 @@ public:
 
 
     int numDependencies = numDependencyBytes / sizeof(uint64_t);
-    if (numDependencyBytes != numDependencies * sizeof(uint64_t)) {
+    if (numDependencyBytes != static_cast<int>(numDependencies * sizeof(uint64_t))) {
       *error_out = (llvm::Twine("unexpected contents for database result: ") +
                     llvm::Twine((int)dbKeyID.value)).str();
       return false;
@@ -767,7 +767,7 @@ public:
       
       // map dependencies
       int numDependencies = numDependencyBytes / sizeof(uint64_t);
-      if (numDependencyBytes != numDependencies * sizeof(uint64_t)) {
+      if (numDependencyBytes != static_cast<int>(numDependencies * sizeof(uint64_t))) {
         *error_out = (llvm::Twine("unexpected contents for database result: ") +
                       llvm::Twine((int)dbKeyID.value)).str();
         return false;
